Fix: add onload hooks
