+++ src/components/App/App.css
@@ -1,3 +1,2 @@
 @import url(../../blocks/mainlink.css);
 @import url(../../blocks/logo.css);
-@import url(../../blocks/main.css);
