+++ src/components/Pages/MainPage/main.css
@@ -1,6 +1,7 @@
 .main {
   width: 100%;
   overflow: hidden;
+  margin-top: 74px;
 }
 
 .main__section {
