+++ src/components/Pages/MoviesPage/MoviesPage.js
@@ -1,10 +1,10 @@
 import PropTypes from 'prop-types';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import ReactPaginate from 'react-paginate';
-import Api from '../../../utils/api';
 import VideoCard from '../../Cards/VideoCard/VideoCard';
 import Heading from '../../UI/Heading/Heading';
 import ScrollContainer from '../../UI/ScrollContainer/ScrollContainer';
+import useMoviesGenres from '../../../hooks/useMoviesGenres';
 import {
   setActiveFilters,
   filterItemByFiltersList,
@@ -13,34 +13,11 @@ import {
 import './MoviesPage.css';
 
 const MoviesPage = ({ handleVideoClick }) => {
-  const [moviesData, setMoviesData] = useState([]);
-  const [genres, setGenres] = useState([]);
   const [activeGenres, setActiveGenres] = useState(['Все']);
   const [currentPage, setCurrentPage] = useState(0);
   const perPage = 16;
   const offset = currentPage * perPage;
-  let pageCount;
-
-  useEffect(() => {
-    Api.getMovies()
-      .then((data) => {
-        const movies = data.map((movie) => ({
-          tagNames: movie.tags.map((tag) => tag.name),
-          ...movie,
-        }));
-        setMoviesData(movies);
-        // Список уникальных жанров для кнопок фильтра-рубрикатора
-
-        const genresData = movies
-          .map((movie) => movie.tagNames)
-          .flat()
-          .filter((item, i, arr) => arr.indexOf(item) === i);
-        setGenres(['Все', ...genresData]);
-        // Подсчёт кол-ва страниц: округление частного общего кол-ва на кол-во на одной странице
-        pageCount = Math.ceil(movies.length / perPage);
-      })
-      .catch((err) => console.log(`Ошибка: ${err}`));
-  }, []);
+  const { moviesData, genres, pageCount } = useMoviesGenres(perPage);
 
   const currentPageData = moviesData
     .slice(offset, offset + perPage)
