+++ src/components/Pages/ReadAndWatch/ReadAndWatch.js
@@ -1,33 +1,10 @@
 import PropTypes from 'prop-types';
-import { useEffect, useState } from 'react';
-import Api from '../../../utils/api';
 import ReadAndWatchSection from '../../Containers/ReadAndWatchSection/ReadAndWatchSection';
+import useReadWatch from '../../../hooks/useReadWatch';
 import './ReadAndWatch.css';
 
 const ReadAndWatch = ({ handleVideoClick }) => {
-  const [guideData, setGuideData] = useState([]);
-  const [videosData, setVideosData] = useState([]);
-  const [articlesData, setArticlesData] = useState([]);
-  const [moviesData, setMoviesData] = useState([]);
-  const [booksData, setBooksData] = useState([]);
-
-  useEffect(() => {
-    Promise.all([
-      Api.getMaterials(),
-      Api.getVideos(),
-      Api.getArticles(),
-      Api.getMovies(),
-      Api.getBooks(),
-    ])
-      .then(([materials, videos, articles, movies, books]) => {
-        setGuideData(materials);
-        setVideosData(videos);
-        setArticlesData(articles);
-        setMoviesData(movies);
-        setBooksData(books);
-      })
-      .catch(console.log);
-  }, []);
+  const { guideData, videosData, articlesData, moviesData, booksData } = useReadWatch();
 
   return (
     <section className="raw content main__section">
