+++ src/hooks/useArticles.js
@@ -1,0 +1,21 @@
+import { useState, useEffect } from 'react';
+import Api from '../utils/api';
+
+export default function useArticles({ perPage }) {
+  const [articlesData, setArticlesData] = useState([]);
+  const [chosenArticle, setChosenArticle] = useState({});
+  let pageCount;
+
+  useEffect(() => {
+    Api.getArticles()
+      .then((data) => {
+        const chosen = data.filter((article) => article.chosen)[0];
+        setChosenArticle({ ...chosen });
+        setArticlesData(data);
+        pageCount = Math.ceil(data.length / perPage);
+      })
+      .catch((err) => console.log(`Ошибка: ${err}`));
+  }, []);
+
+  return { articlesData, chosenArticle, pageCount };
+}
